@@ -1,7 +1,7 @@
 [package]
 name = "min-timespan-delivery"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 atomic_float = "1.1.0"
